feat: Connect Redux to fake API (AsyncStorage)
