+++ app/server/api.js
@@ -1,5 +1,8 @@
 import { AsyncStorage } from 'react-native';
-import { CALENDAR_STORAGE_KEY } from './_calendar';
+import { CALENDAR_STORAGE_KEY, formatCalendarResults } from './_calendar';
+
+export const fetchCalendarResults = () =>
+  AsyncStorage.getItem(CALENDAR_STORAGE_KEY).then(formatCalendarResults);
 
 export const submitEntry = (key, entry) =>
   AsyncStorage.mergeItem(
