+++ app/state/utils/helpers.js
@@ -1,0 +1,3 @@
+export const getDailyReminderValue = () => ({
+  today: "👋 Don't forget to log your data today!"
+});
