+++ app/state/utils/index.js
@@ -1,1 +1,4 @@
+import * as helpers from './helpers';
+
 export { default as createReducer } from './createReducer';
+export { helpers };
